+++ js/main.js
@@ -1,10 +1,11 @@
 import {renderThumbnails} from './thumbnails.js';
 import {getPhotos} from './api.js';
 import {debounce, showAlert} from './util.js';
-import './form.js';
-import './image-effects.js';
 import {setSwitchFiltersView, showFilters} from './filters.js';
+import {hideSlider} from './image-effects.js';
+import './form.js';
 
+hideSlider();
 getPhotos((photos) => {
   renderThumbnails(photos);
   setSwitchFiltersView(photos, debounce(renderThumbnails));
